API

API

Types

RadiationSpectra.AbstractFitFunction

RadiationSpectra.FitFunction

Functions

RadiationSpectra.calibrate_spectrum

RadiationSpectra.gauss

RadiationSpectra.gauss

RadiationSpectra.get_example_spectrum

RadiationSpectra.llhfit!

RadiationSpectra.lsqfit!

RadiationSpectra.peakfinder

Documentation

RadiationSpectra.peakfinder — Method.

peakfinder(h::Histogram; <keyword arguments>)::Tuple{Histogram, Array{Float64, 1}}

Returns a deconvoluted spectrum and an array of peak positions.

Keywords

sigma::Real=2.0: The expected sigma of a peak in the spectrum.

threshold::Real=10.0: Threshold for being identified as a peak in the deconvoluted spectrum.

backgroundRemove::Bool=true

deconIterations::Int=3

markov::Bool=true

averWindow::Int=3

Source

This function is basically a copy of TSpectrum::SearchHighRes from ROOT.

M.A. Mariscotti: A method for identification of peaks in the presence of background and its application to spectrum analysis. NIM 50 (1967), 309-320.

M. Morhac;, J. Kliman, V. Matouoek, M. Veselsky, I. Turzo.:Identification of peaks in multidimensional coincidence gamma-ray spectra. NIM, A443 (2000) 108-125.

Z.K. Silagadze, A new algorithm for automatic photopeak searches. NIM A 376 (1996), 451.

source

RadiationSpectra.AbstractFitFunction — Type.

AbstractFitFunction{T, N}

Abstract type for an N-dimensional fit of eltype T.

source

RadiationSpectra.FitFunction — Type.

FitFunction{T} <: AbstractFitFunction{T, 1}

Fields:

model::Function: Function of the fit model.

fitrange::Tuple{Union{Missing, T}, Union{Missing, T}}: Range on which the fit is performed.

parameters::Union{Vector{Missing}, Vector{T}}: Fitted parameters.

uncertainties::Union{Vector{Missing}, Vector{T}}: Uncertainties of the fitted parameters.

initial_parameters::Union{Vector{Missing}, Vector{T}}: Initial parameters.

confidence_level::Union{Missing, T}: Confidence level, used to determine the uncertainties.

Plotting:

A plot recipes exists for this struct: plot(fit::FitFunction{T}; npoints=501)

Plots the model function over the fitrange with 501-points.

source

RadiationSpectra.calibrate_spectrum — Method.

calibrate_spectrum(h_uncal::Histogram, photon_lines::Array{Real, 1})::Histogram

Calibrate the spectrum h_uncal. This is done by:

finding peaks through devoncolution

identifying them through comparison of the ratios of their positions with the ratios of the known lines

fitting all identified peaks (with a gaussian plus first order polynomial) to get their position more precisely

performe a linear fit (offset forced to 0) of these positions vs the true positions (lines) to get the calibration constant

source

RadiationSpectra.gauss — Method.

gauss(x::Vector{T}, p::Vector{T})::Vector{T}

Maps x to gauss(x::T, p::Vector{T})::T

source

RadiationSpectra.gauss — Method.

gauss(x::T, p::Vector{T})::T

A Gaussian with 3 parameters:

p[1]: Scale/Amplitude

p[2]: σ

p[3]: μ

source

RadiationSpectra.get_example_spectrum — Method.

get_example_spectrum()::Histogram

Returns an uncalibrated radiation spectrum for testing and demonstrating purpose.

source

RadiationSpectra.llhfit! — Method.

 llhfit!(fit::FitFunction, h::Histogram)::Nothing

Performs a log-likelihood fit of the model function fit.model and the initial parameters fit.initial_parameters on the histogram h in the range fit.fitrange. The determined parameters are stored in fit.parameters.

The likelihood for each individual bin is the Poission distribution.

There are no uncertainty estimations for this fit yet.

source

RadiationSpectra.lsqfit! — Method.

lsqfit!(fit::FitFunction, h::Histogram)::Nothing

Performs a Least Square Fit with the model fit.model and the initial parameters fit.initial_parameters on the histogram h in the range fit.fitrange. The determined parameters are stored in fit.parameters and the corresponding uncertainties in fit.uncertainties for the given confidence level fit.confidence_level.

The uncertainties are marginalizations of the covariance matrix determined by the LSQFit.jl package.

See LsqFit.jl for more detail.

source